server/requirements.txt
praat-parselmouth
transformers
matplotlib
opencv-python
requests
pandas
numpy
sounddevice
pydub
mediapipe
moviepy
ffmpeg-python